move sendmessages code
